web-server/requirements.txt
uvicorn[standard]==0.20.0
gunicorn==20.1.0
fastapi[all]==0.104.1
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
alembic==1.11.2
pika==1.3.2
requests==2.31.0
cloudscraper==1.2.71
lxml==4.9.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
